Refactor HashType to its own module.
