+++ src/poseidon.rs
@@ -1,3 +1,4 @@
+use crate::hash_type::HashType;
 use crate::matrix::Matrix;
 use crate::mds::{create_mds_matrices, factor_to_sparse_matrixes, MDSMatrices, SparseMatrix};
 use crate::poseidon_alt::{hash_correct, hash_optimized_dynamic};
@@ -11,94 +12,6 @@ use paired::bls12_381::Bls12;
 use std::marker::PhantomData;
 use typenum::marker_traits::Unsigned;
 use typenum::*;
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum HashType<Fr: PrimeField, A: Arity<Fr>> {
-    MerkleTree,
-    MerkleTreeSparse(u64),
-    VariableLength,
-    ConstantLength(usize),
-    Encryption,
-    Custom(CType<Fr, A>),
-}
-
-impl<Fr: PrimeField, A: Arity<Fr>> HashType<Fr, A> {
-    fn capacity_tag(&self, strength: &Strength) -> Fr {
-        let pow2 = |n| pow2::<Fr, A>(n);
-        let x_pow2 = |coeff, n| x_pow2::<Fr, A>(coeff, n);
-        let with_strength = |x: Fr| {
-            let mut tmp = x;
-            tmp.add_assign(&Self::strength_tag_component(strength));
-            tmp
-        };
-
-        match self {
-            // 2^arity - 1
-            HashType::MerkleTree => with_strength(arity_tag::<Fr, A>()),
-            // bitmask
-            HashType::MerkleTreeSparse(bitmask) => with_strength(scalar_from_u64(*bitmask)),
-            // 2^64
-            HashType::VariableLength => with_strength(pow2(64)),
-            // length * 2^64
-            HashType::ConstantLength(length) => {
-                assert!(*length as usize <= A::to_usize());
-                with_strength(x_pow2(*length as u64, 64))
-            }
-            // 2^32
-            HashType::Encryption => with_strength(pow2(32)),
-            // identifier * 2^32
-            HashType::Custom(ref ctype) => ctype.capacity_tag(&strength),
-        }
-    }
-
-    fn strength_tag_component(strength: &Strength) -> Fr {
-        let id = match strength {
-            // Standard strength doesn't affect the base tag.
-            Strength::Standard => 0,
-            Strength::Strengthened => 1,
-        };
-
-        x_pow2::<Fr, A>(id, 32)
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum CType<Fr: PrimeField, A: Arity<Fr>> {
-    Arbitrary(u64),
-    _Phantom((Fr, A)),
-}
-
-impl<Fr: PrimeField, A: Arity<Fr>> CType<Fr, A> {
-    fn identifier(&self) -> u64 {
-        match self {
-            CType::Arbitrary(id) => *id,
-            CType::_Phantom(_) => panic!("_Phantom is not a real custom tag type."),
-        }
-    }
-
-    fn capacity_tag(&self, _strength: &Strength) -> Fr {
-        x_pow2::<Fr, A>(self.identifier(), 32)
-    }
-}
-
-/// pow2(n) = 2^n
-fn pow2<Fr: PrimeField, A: Arity<Fr>>(n: i32) -> Fr {
-    let two: Fr = scalar_from_u64(2);
-    two.pow([n as u64, 0, 0, 0])
-}
-
-/// x_pow2(x, n) = x * 2^n
-fn x_pow2<Fr: PrimeField, A: Arity<Fr>>(coeff: u64, n: i32) -> Fr {
-    let mut tmp: Fr = pow2::<Fr, A>(n);
-    tmp.mul_assign(&scalar_from_u64(coeff));
-    tmp
-}
-
-/// The arity tag is the first element of a Poseidon permutation.
-/// This extra element is necessary for 128-bit security.
-pub fn arity_tag<Fr: PrimeField, A: Arity<Fr>>() -> Fr {
-    A::tag()
-}
 
 /// Available arities for the Poseidon hasher.
 pub trait Arity<T>: ArrayLength<T> {
@@ -197,6 +110,8 @@ where
     pub pre_sparse_matrix: Matrix<E::Fr>,
     pub sparse_matrixes: Vec<SparseMatrix<E>>,
     pub strength: Strength,
+    /// The domain tag is the first element of a Poseidon permutation.
+    /// This extra element is necessary for 128-bit security.
     pub domain_tag: E::Fr,
     pub full_rounds: usize,
     pub half_full_rounds: usize,
@@ -246,7 +161,7 @@ where
             _ => panic!("cannot set constant length of hash without type ConstantLength."),
         };
 
-        let domain_tag = hash_type.capacity_tag(&self.strength);
+        let domain_tag = hash_type.domain_tag(&self.strength);
 
         Self {
             hash_type,
@@ -298,7 +213,7 @@ where
             pre_sparse_matrix,
             sparse_matrixes,
             strength,
-            domain_tag: hash_type.capacity_tag(&strength),
+            domain_tag: hash_type.domain_tag(&strength),
             full_rounds,
             half_full_rounds,
             partial_rounds,
